feat: create separate collector profile mutation
Current mutation is hitting /me/collector_profile and we need the ability to update a profile when not in a /me/ context.

This commit creates a separate mutation that can be called with a profile ID.
