+++ _schemaV2.graphql
@@ -11016,6 +11016,11 @@ type Mutation {
   updateCollectorProfile(
     input: UpdateCollectorProfileInput!
   ): UpdateCollectorProfilePayload
+
+  # Updating a collector profile (loyalty applicant status).
+  updateCollectorProfileWithID(
+    input: UpdateCollectorProfileWithIDInput!
+  ): UpdateCollectorProfileWithIDPayload
   updateConsignmentSubmission(
     # Parameters for UpdateSubmissionMutation
     input: UpdateSubmissionMutationInput!
@@ -15462,6 +15467,83 @@ input UpdateCollectorProfileInput {
 }
 
 type UpdateCollectorProfilePayload {
+  clientMutationId: String
+  collectorLevel: Int
+  companyName: String
+  companyWebsite: String
+  confirmedBuyerAt(
+    format: String
+
+    # A tz database time zone, otherwise falls back to "X-TIMEZONE" header. See
+    # http://www.iana.org/time-zones,
+    # https://en.wikipedia.org/wiki/List_of_tz_database_time_zones
+    timezone: String
+  ): String
+  email: String
+
+  # A globally unique ID.
+  id: ID!
+  institutionalAffiliations: String
+  intents: [String]
+
+  # A type-specific ID likely used as a database ID.
+  internalID: ID!
+  loyaltyApplicantAt(
+    format: String
+
+    # A tz database time zone, otherwise falls back to "X-TIMEZONE" header. See
+    # http://www.iana.org/time-zones,
+    # https://en.wikipedia.org/wiki/List_of_tz_database_time_zones
+    timezone: String
+  ): String
+  name: String
+  privacy: String
+  professionalBuyerAppliedAt(
+    format: String
+
+    # A tz database time zone, otherwise falls back to "X-TIMEZONE" header. See
+    # http://www.iana.org/time-zones,
+    # https://en.wikipedia.org/wiki/List_of_tz_database_time_zones
+    timezone: String
+  ): String
+  professionalBuyerAt(
+    format: String
+
+    # A tz database time zone, otherwise falls back to "X-TIMEZONE" header. See
+    # http://www.iana.org/time-zones,
+    # https://en.wikipedia.org/wiki/List_of_tz_database_time_zones
+    timezone: String
+  ): String
+  selfReportedPurchases: String
+  userInterests: [UserInterest]!
+}
+
+input UpdateCollectorProfileWithIDInput {
+  # List of affiliated auction house ids, referencing Galaxy.
+  affiliatedAuctionHouseIds: [String]
+
+  # List of affiliated fair ids, referencing Galaxy.
+  affiliatedFairIds: [String]
+
+  # List of affiliated gallery ids, referencing Galaxy.
+  affiliatedGalleryIds: [String]
+  clientMutationId: String
+  companyName: String
+  companyWebsite: String
+  confirmedBuyer: Boolean
+
+  # The internal ID of the collector profile to update
+  id: String
+  institutionalAffiliations: String
+  intents: [Intents]
+  loyaltyApplicant: Boolean
+  professionalBuyer: Boolean
+
+  # Free-form text of purchases the collector has indicated.
+  selfReportedPurchases: String
+}
+
+type UpdateCollectorProfileWithIDPayload {
   clientMutationId: String
   collectorLevel: Int
   companyName: String
