+++ src/schema/v2/CollectorProfile/collectorProfile.ts
@@ -1,0 +1,60 @@
+import date from "schema/v2/fields/date"
+import { InternalIDFields } from "schema/v2/object_identification"
+import {
+  GraphQLObjectType,
+  GraphQLString,
+  GraphQLInt,
+  GraphQLList,
+  GraphQLFieldConfigMap,
+  GraphQLNonNull,
+} from "graphql"
+import { ResolverContext } from "types/graphql"
+import { userInterestType } from "../me/userInterests"
+
+export const CollectorProfileFields: GraphQLFieldConfigMap<
+  any,
+  ResolverContext
+> = {
+  ...InternalIDFields,
+  collectorLevel: {
+    type: GraphQLInt,
+    resolve: ({ collector_level }) => collector_level,
+  },
+  companyName: {
+    type: GraphQLString,
+    resolve: ({ company_name }) => company_name,
+  },
+  companyWebsite: {
+    type: GraphQLString,
+    resolve: ({ company_website }) => company_website,
+  },
+  confirmedBuyerAt: date,
+  email: { type: GraphQLString },
+  institutionalAffiliations: {
+    type: GraphQLString,
+    resolve: ({ institutional_affiliations }) => institutional_affiliations,
+  },
+  intents: { type: new GraphQLList(GraphQLString) },
+  loyaltyApplicantAt: date,
+  name: { type: GraphQLString },
+  privacy: { type: GraphQLString },
+  professionalBuyerAppliedAt: date,
+  professionalBuyerAt: date,
+  selfReportedPurchases: {
+    type: GraphQLString,
+    resolve: ({ self_reported_purchases }) => self_reported_purchases,
+  },
+  userInterests: {
+    type: new GraphQLNonNull(new GraphQLList(userInterestType)),
+    resolve: (_collectorProfile, _args, { userInterestsLoader }) => {
+      return userInterestsLoader?.()
+    },
+  },
+}
+
+export const CollectorProfileType = new GraphQLObjectType<any, ResolverContext>(
+  {
+    name: "CollectorProfileType",
+    fields: CollectorProfileFields,
+  }
+)
