+++ src/schema/v2/me/collector_profile.ts
@@ -1,64 +1,6 @@
-import date from "schema/v2/fields/date"
-import { InternalIDFields } from "schema/v2/object_identification"
-import {
-  GraphQLObjectType,
-  GraphQLString,
-  GraphQLInt,
-  GraphQLList,
-  GraphQLFieldConfig,
-  GraphQLFieldConfigMap,
-  GraphQLNonNull,
-} from "graphql"
+import { GraphQLFieldConfig } from "graphql"
 import { ResolverContext } from "types/graphql"
-import { userInterestType } from "./userInterests"
-
-export const CollectorProfileFields: GraphQLFieldConfigMap<
-  any,
-  ResolverContext
-> = {
-  ...InternalIDFields,
-  collectorLevel: {
-    type: GraphQLInt,
-    resolve: ({ collector_level }) => collector_level,
-  },
-  companyName: {
-    type: GraphQLString,
-    resolve: ({ company_name }) => company_name,
-  },
-  companyWebsite: {
-    type: GraphQLString,
-    resolve: ({ company_website }) => company_website,
-  },
-  confirmedBuyerAt: date,
-  email: { type: GraphQLString },
-  institutionalAffiliations: {
-    type: GraphQLString,
-    resolve: ({ institutional_affiliations }) => institutional_affiliations,
-  },
-  intents: { type: new GraphQLList(GraphQLString) },
-  loyaltyApplicantAt: date,
-  name: { type: GraphQLString },
-  privacy: { type: GraphQLString },
-  professionalBuyerAppliedAt: date,
-  professionalBuyerAt: date,
-  selfReportedPurchases: {
-    type: GraphQLString,
-    resolve: ({ self_reported_purchases }) => self_reported_purchases,
-  },
-  userInterests: {
-    type: new GraphQLNonNull(new GraphQLList(userInterestType)),
-    resolve: (_collectorProfile, _args, { userInterestsLoader }) => {
-      return userInterestsLoader?.()
-    },
-  },
-}
-
-export const CollectorProfileType = new GraphQLObjectType<any, ResolverContext>(
-  {
-    name: "CollectorProfileType",
-    fields: CollectorProfileFields,
-  }
-)
+import { CollectorProfileType } from "../CollectorProfile/collectorProfile"
 
 const CollectorProfile: GraphQLFieldConfig<void, ResolverContext> = {
   type: CollectorProfileType,
