+++ src/schema/v2/schema.ts
@@ -77,6 +77,7 @@ import { myCollectionUpdateArtworkMutation } from "./me/myCollectionUpdateArtwor
 import saveArtworkMutation from "./me/saveArtworkMutation"
 import { sendConfirmationEmailMutation } from "./me/sendConfirmationEmailMutation"
 import UpdateCollectorProfile from "./me/update_collector_profile"
+import UpdateCollectorProfileWithID from "./CollectorProfile/mutations/updateCollectorProfileMutation"
 import UpdateMyUserProfileMutation from "./me/update_me_mutation"
 import { updateMyPasswordMutation } from "./me/updateMyPasswordMutation"
 import { updateUserMutation } from "./users/updateUserMutation"
@@ -313,6 +314,7 @@ export default new GraphQLSchema({
       startIdentityVerification: startIdentityVerificationMutation,
       unlinkAuthentication: unlinkAuthenticationMutation,
       updateCollectorProfile: UpdateCollectorProfile,
+      updateCollectorProfileWithID: UpdateCollectorProfileWithID,
       updateConversation: UpdateConversationMutation,
       updateMyPassword: updateMyPasswordMutation,
       updateUser: updateUserMutation,
